app/requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
dj-database-url==2.1.0
Django==4.2.9
django-cors-headers==4.3.1
django-ratelimit==4.1.0
django-vite==3.0.2
djangorestframework==3.14.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inertia-django==0.5.3
mysql-connector-python==8.3.0
mysqlclient==2.2.1
packaging==23.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
urllib3==2.1.0
zope.interface==6.1
